+++ packages/tracing/src/browser/metrics.ts
@@ -5,7 +5,7 @@ import { getGlobalObject, logger } from '@sentry/utils';
 
 import { Span } from '../span';
 import { Transaction } from '../transaction';
-import { msToSec } from './utils';
+import { msToSec } from '../utils';
 
 const global = getGlobalObject<Window>();
