+++ packages/tracing/src/utils.ts
@@ -1,4 +1,5 @@
-import { Options, TraceparentData } from '@sentry/types';
+import { getCurrentHub, Hub } from '@sentry/hub';
+import { Options, TraceparentData, Transaction } from '@sentry/types';
 
 export const TRACEPARENT_REGEXP = new RegExp(
   '^[ \\t]*' + // whitespace
@@ -40,6 +41,34 @@ export function extractTraceparentData(traceparent: string): TraceparentData | u
     };
   }
   return undefined;
+}
+
+/** Grabs active transaction off scope */
+export function getActiveTransaction<T extends Transaction>(hub: Hub = getCurrentHub()): T | undefined {
+  if (hub) {
+    const scope = hub.getScope();
+    if (scope) {
+      return scope.getTransaction() as T | undefined;
+    }
+  }
+
+  return undefined;
+}
+
+/**
+ * Converts from milliseconds to seconds
+ * @param time time in ms
+ */
+export function msToSec(time: number): number {
+  return time / 1000;
+}
+
+/**
+ * Converts from seconds to milliseconds
+ * @param time time in seconds
+ */
+export function secToMs(time: number): number {
+  return time * 1000;
 }
 
 // so it can be used in manual instrumentation without necessitating a hard dependency on @sentry/utils
